Checking for file open errors

+++ intel8x0.c
@@ -111,7 +111,7 @@ static struct file* pcmFile;
 struct file* file_open(const char* path, int flags, int rights) {
     struct file* filp = NULL;
     mm_segment_t oldfs;
-    int err = 0;
+    long err = 0;
 
     oldfs = get_fs();
     set_fs(get_ds());
